[ALPS03223821] DISP: fix coverity
[Detail]
Fix coverity:
  Missing break in switch:1999320
  Bad bit shift operation:5459460
  Dereference after null check:8259973

MTK-Commit-Id: aa1bc333bede75d31da41b257c98c9c1851eae64

CR-Id: ALPS03223821
Feature: [Module]Display Driver

+++ drivers/misc/mediatek/video/mt6771/dispsys/ddp_ovl.h
@@ -60,7 +60,7 @@ unsigned int ovl_to_index(enum DISP_MODULE_ENUM module);
 void ovl_get_info(enum DISP_MODULE_ENUM module, void *data);
 unsigned int ddp_ovl_get_cur_addr(bool rdma_mode, int layerid);
 enum DISP_MODULE_ENUM ovl_index_to_mod_for_debug(int index);
-unsigned long ovl_layer_num_for_debug(enum DISP_MODULE_ENUM module);
+int ovl_layer_num_for_debug(enum DISP_MODULE_ENUM module);
 
 void ovl_set_dbg_mode(bool mode);
 unsigned int ovl_set_bg_color(unsigned int bg_color);
